+++ src/plugins/data_source_management/public/components/edit_data_source/components/edit_form/edit_data_source_form.tsx
@@ -29,6 +29,7 @@ import {
   credentialSourceOptions,
   DataSourceAttributes,
   DataSourceManagementContextValue,
+  SigV4Content,
   ToastMessageItem,
   UsernamePasswordTypedContent,
 } from '../../../../types';
@@ -40,6 +41,7 @@ import {
   performDataSourceFormValidation,
 } from '../../../validation';
 import { UpdatePasswordModal } from '../update_password_modal';
+import { UpdateAwsCredentialModal } from '../update_aws_credential_modal';
 
 export interface EditDataSourceProps {
   existingDataSource: DataSourceAttributes;
@@ -56,9 +58,10 @@ export interface EditDataSourceState {
   endpoint: string;
   auth: {
     type: AuthType;
-    credentials: UsernamePasswordTypedContent;
+    credentials: UsernamePasswordTypedContent | SigV4Content | undefined;
   };
   showUpdatePasswordModal: boolean;
+  showUpdateAwsCredentialModal: boolean;
   showUpdateOptions: boolean;
   isLoading: boolean;
 }
@@ -81,9 +84,13 @@ export class EditDataSourceForm extends React.Component<EditDataSourceProps, Edi
         credentials: {
           username: '',
           password: '',
+          region: '',
+          accessKey: '',
+          secretKey: '',
         },
       },
       showUpdatePasswordModal: false,
+      showUpdateAwsCredentialModal: false,
       showUpdateOptions: false,
       isLoading: false,
     };
@@ -102,6 +109,11 @@ export class EditDataSourceForm extends React.Component<EditDataSourceProps, Edi
     if (this.props.existingDataSource) {
       const { title, description, endpoint, auth } = this.props.existingDataSource;
 
+      const authTypeCheckResults = {
+        isUserNamePassword: auth.type === AuthType.UsernamePasswordType,
+        isSigV4: auth.type === AuthType.SigV4,
+      };
+
       this.setState({
         title,
         description: description || '',
@@ -109,8 +121,11 @@ export class EditDataSourceForm extends React.Component<EditDataSourceProps, Edi
         auth: {
           type: auth.type,
           credentials: {
-            username: auth.type === AuthType.NoAuth ? '' : auth.credentials?.username || '',
-            password: auth.type === AuthType.NoAuth ? '' : this.maskedPassword,
+            username: authTypeCheckResults.isUserNamePassword ? auth.credentials?.username : '',
+            password: authTypeCheckResults.isUserNamePassword ? this.maskedPassword : '',
+            region: authTypeCheckResults.isSigV4 ? auth.credentials?.region : '',
+            accessKey: authTypeCheckResults.isSigV4 ? this.maskedPassword : '',
+            secretKey: authTypeCheckResults.isSigV4 ? this.maskedPassword : '',
           },
         },
       });
@@ -148,20 +163,7 @@ export class EditDataSourceForm extends React.Component<EditDataSourceProps, Edi
   };
 
   onChangeAuthType = (value: string) => {
-    const valueToSave =
-      value === AuthType.UsernamePasswordType ? AuthType.UsernamePasswordType : AuthType.NoAuth;
-
-    const formErrorsByField = {
-      ...this.state.formErrorsByField,
-      createCredential: { ...this.state.formErrorsByField.createCredential },
-    };
-    if (valueToSave === AuthType.NoAuth) {
-      formErrorsByField.createCredential = {
-        username: [],
-        password: [],
-      };
-    }
-    this.setState({ auth: { ...this.state.auth, type: valueToSave }, formErrorsByField }, () => {
+    this.setState({ auth: { ...this.state.auth, type: value as AuthType } }, () => {
       this.onChangeFormValues();
     });
   };
@@ -174,7 +176,10 @@ export class EditDataSourceForm extends React.Component<EditDataSourceProps, Edi
     this.setState({
       auth: {
         ...this.state.auth,
-        credentials: { ...this.state.auth.credentials, username: e.target.value },
+        credentials: {
+          ...this.state.auth.credentials,
+          username: e.target.value,
+        } as UsernamePasswordTypedContent,
       },
     });
   };
@@ -208,7 +213,76 @@ export class EditDataSourceForm extends React.Component<EditDataSourceProps, Edi
     this.setState({
       auth: {
         ...this.state.auth,
-        credentials: { ...this.state.auth.credentials, password: e.target.value },
+        credentials: {
+          ...this.state.auth.credentials,
+          password: e.target.value,
+        } as UsernamePasswordTypedContent,
+      },
+    });
+  };
+
+  onChangeRegion = (e: { target: { value: any } }) => {
+    this.setState({
+      auth: {
+        ...this.state.auth,
+        credentials: { ...this.state.auth.credentials, region: e.target.value } as SigV4Content,
+      },
+    });
+  };
+
+  validateRegion = () => {
+    const isValid = !!this.state.auth.credentials.region?.trim().length;
+    this.setState({
+      formErrorsByField: {
+        ...this.state.formErrorsByField,
+        awsCredential: {
+          ...this.state.formErrorsByField.awsCredential,
+          region: isValid ? [] : [''],
+        },
+      },
+    });
+  };
+
+  onChangeAccessKey = (e: { target: { value: any } }) => {
+    this.setState({
+      auth: {
+        ...this.state.auth,
+        credentials: { ...this.state.auth.credentials, accessKey: e.target.value } as SigV4Content,
+      },
+    });
+  };
+
+  validateAccessKey = () => {
+    const isValid = !!this.state.auth.credentials.accessKey;
+    this.setState({
+      formErrorsByField: {
+        ...this.state.formErrorsByField,
+        awsCredential: {
+          ...this.state.formErrorsByField.awsCredential,
+          accessKey: isValid ? [] : [''],
+        },
+      },
+    });
+  };
+
+  onChangeSecretKey = (e: { target: { value: any } }) => {
+    this.setState({
+      auth: {
+        ...this.state.auth,
+        credentials: { ...this.state.auth.credentials, secretKey: e.target.value } as SigV4Content,
+      },
+    });
+  };
+
+  validateSecretKey = () => {
+    const isValid = !!this.state.auth.credentials.secretKey;
+    this.setState({
+      formErrorsByField: {
+        ...this.state.formErrorsByField,
+        awsCredential: {
+          ...this.state.formErrorsByField.awsCredential,
+          secretKey: isValid ? [] : [''],
+        },
       },
     });
   };
@@ -221,12 +295,30 @@ export class EditDataSourceForm extends React.Component<EditDataSourceProps, Edi
         description: this.state.description,
         auth: this.state.auth,
       };
-      /* Remove credentials object for NoAuth */
-      if (this.state.auth.type === AuthType.NoAuth) {
-        delete formValues.auth.credentials;
-      } else if (this.props.existingDataSource.auth.type === AuthType.UsernamePasswordType) {
-        /* Remove password if previously & currently username & password method is selected*/
-        delete formValues.auth.credentials?.password;
+
+      switch (this.state.auth.type) {
+        case AuthType.NoAuth:
+          delete formValues.auth.credentials;
+          break;
+        case AuthType.SigV4:
+          delete formValues.auth.credentials?.username;
+          delete formValues.auth.credentials?.password;
+          /* Remove access key and secret key if previously & currently SigV4 auth method is selected*/
+          if (this.props.existingDataSource.auth.type === this.state.auth.type) {
+            delete formValues.auth.credentials?.accessKey;
+            delete formValues.auth.credentials?.secretKey;
+          }
+          break;
+        case AuthType.UsernamePasswordType:
+          delete formValues.auth.credentials?.accessKey;
+          delete formValues.auth.credentials?.secretKey;
+          delete formValues.auth.credentials?.region;
+          /* Remove password if previously & currently username & password auth method is selected*/
+          if (this.props.existingDataSource.auth.type === this.state.auth.type)
+            delete formValues.auth.credentials?.password;
+          break;
+        default:
+          break;
       }
 
       /* Submit */
@@ -303,6 +395,10 @@ export class EditDataSourceForm extends React.Component<EditDataSourceProps, Edi
     this.setState({ showUpdatePasswordModal: true });
   };
 
+  onClickUpdateAwsCredential = () => {
+    this.setState({ showUpdateAwsCredentialModal: true });
+  };
+
   /* Update password */
   updatePassword = async (password: string) => {
     const { title, description, auth } = this.props.existingDataSource;
@@ -315,7 +411,7 @@ export class EditDataSourceForm extends React.Component<EditDataSourceProps, Edi
         credentials: {
           username: auth.credentials ? auth.credentials.username : '',
           password,
-        },
+        } as UsernamePasswordTypedContent,
       },
     };
     this.closePasswordModal();
@@ -335,11 +431,48 @@ export class EditDataSourceForm extends React.Component<EditDataSourceProps, Edi
     }
   };
 
+  /* Update aws credential */
+  updateAwsCredential = async (accessKey: string, secretKey: string) => {
+    const { title, description, auth } = this.props.existingDataSource;
+    const updateAttributes: DataSourceAttributes = {
+      title,
+      description,
+      endpoint: undefined,
+      auth: {
+        type: auth.type,
+        credentials: {
+          region: auth.credentials ? auth.credentials.region : '',
+          accessKey,
+          secretKey,
+        } as SigV4Content,
+      },
+    };
+    this.closeAwsCredentialModal();
+
+    try {
+      await this.props.handleSubmit(updateAttributes);
+      this.props.displayToastMessage({
+        id: 'dataSourcesManagement.editDataSource.updatePasswordSuccessMsg',
+        defaultMessage: 'Password updated successfully.',
+        success: true,
+      });
+    } catch (e) {
+      this.props.displayToastMessage({
+        id: 'dataSourcesManagement.editDataSource.updatePasswordFailMsg',
+        defaultMessage: 'Updating the stored password failed with some errors.',
+      });
+    }
+  };
+
   /* Render methods */
 
-  /* Render Modal for new credential */
+  /* Render modal for new credential */
   closePasswordModal = () => {
     this.setState({ showUpdatePasswordModal: false });
+  };
+
+  closeAwsCredentialModal = () => {
+    this.setState({ showUpdateAwsCredentialModal: false });
   };
 
   renderUpdatePasswordModal = () => {
@@ -367,6 +500,33 @@ export class EditDataSourceForm extends React.Component<EditDataSourceProps, Edi
       </>
     );
   };
+
+  renderUpdateAwsCredentialModal = () => {
+    return (
+      <>
+        <EuiButton
+          onClick={this.onClickUpdateAwsCredential}
+          data-test-subj="editDatasourceUpdateAwsCredentialBtn"
+        >
+          {
+            <FormattedMessage
+              id="dataSourcesManagement.editDataSource.updateStoredAwsCredential"
+              defaultMessage="Update stored AWS credential"
+            />
+          }
+        </EuiButton>
+
+        {this.state.showUpdateAwsCredentialModal ? (
+          <UpdateAwsCredentialModal
+            region={this.state.auth?.credentials?.region || ''}
+            handleUpdateAwsCredential={this.updateAwsCredential}
+            closeUpdateAwsCredentialModal={this.closeAwsCredentialModal}
+          />
+        ) : null}
+      </>
+    );
+  };
+
   /* Render header*/
   renderHeader = () => {
     return (
@@ -575,8 +735,106 @@ export class EditDataSourceForm extends React.Component<EditDataSourceProps, Edi
         </EuiFormRow>
 
         <EuiSpacer />
+        {this.renderSelectedAuthType(this.state.auth.type)}
+      </>
+    );
+  };
 
-        {this.state.auth.type !== AuthType.NoAuth ? this.renderUsernamePasswordFields() : null}
+  renderSelectedAuthType = (type: AuthType) => {
+    switch (type) {
+      case AuthType.UsernamePasswordType:
+        return this.renderUsernamePasswordFields();
+      case AuthType.SigV4:
+        return this.renderSigV4ContentFields();
+      default:
+        return null;
+    }
+  };
+
+  renderSigV4ContentFields = () => {
+    return (
+      <>
+        <EuiFormRow
+          label={i18n.translate('dataSourcesManagement.createDataSource.region', {
+            defaultMessage: 'Region',
+          })}
+          isInvalid={!!this.state.formErrorsByField.awsCredential.region.length}
+          error={this.state.formErrorsByField.awsCredential.region}
+        >
+          <EuiFieldText
+            placeholder={i18n.translate(
+              'dataSourcesManagement.createDataSource.regionPlaceholder',
+              {
+                defaultMessage: 'AWS Region, e.g. us-west-2',
+              }
+            )}
+            isInvalid={!!this.state.formErrorsByField.awsCredential.region.length}
+            value={this.state.auth.credentials?.region || ''}
+            onChange={this.onChangeRegion}
+            onBlur={this.validateRegion}
+            data-test-subj="updateDataSourceFormRegionField"
+          />
+        </EuiFormRow>
+        <EuiFormRow
+          label={i18n.translate('dataSourcesManagement.createDataSource.accessKey', {
+            defaultMessage: 'Access Key',
+          })}
+          isInvalid={!!this.state.formErrorsByField.awsCredential.accessKey.length}
+          error={this.state.formErrorsByField.awsCredential.accessKey}
+        >
+          <EuiFieldPassword
+            isInvalid={!!this.state.formErrorsByField.awsCredential.accessKey.length}
+            placeholder={i18n.translate(
+              'dataSourcesManagement.createDataSource.accessKeyPlaceholder',
+              {
+                defaultMessage: 'AWS access key',
+              }
+            )}
+            type={'dual'}
+            value={
+              this.props.existingDataSource.auth.type === AuthType.SigV4
+                ? this.maskedPassword
+                : this.state.auth.credentials?.accessKey
+            }
+            onChange={this.onChangeAccessKey}
+            onBlur={this.validateAccessKey}
+            spellCheck={false}
+            disabled={this.props.existingDataSource.auth.type === AuthType.SigV4}
+            data-test-subj="updateDataSourceFormAccessKeyField"
+          />
+        </EuiFormRow>
+        <EuiFormRow
+          label={i18n.translate('dataSourcesManagement.createDataSource.secretKey', {
+            defaultMessage: 'Secret Key',
+          })}
+          isInvalid={!!this.state.formErrorsByField.awsCredential.secretKey.length}
+          error={this.state.formErrorsByField.awsCredential.secretKey}
+        >
+          <EuiFieldPassword
+            isInvalid={!!this.state.formErrorsByField.awsCredential.secretKey.length}
+            placeholder={i18n.translate(
+              'dataSourcesManagement.createDataSource.secretKeyPlaceholder',
+              {
+                defaultMessage: 'AWS secret key',
+              }
+            )}
+            type={'dual'}
+            value={
+              this.props.existingDataSource.auth.type === AuthType.SigV4
+                ? this.maskedPassword
+                : this.state.auth.credentials?.secretKey
+            }
+            onChange={this.onChangeSecretKey}
+            onBlur={this.validateSecretKey}
+            spellCheck={false}
+            disabled={this.props.existingDataSource.auth.type === AuthType.SigV4}
+            data-test-subj="updateDataSourceFormSecretKeyField"
+          />
+        </EuiFormRow>
+        <EuiSpacer />
+        {this.props.existingDataSource.auth.type === AuthType.SigV4
+          ? this.renderUpdateAwsCredentialModal()
+          : null}
       </>
     );
   };
@@ -600,7 +858,7 @@ export class EditDataSourceForm extends React.Component<EditDataSourceProps, Edi
                 defaultMessage: 'Username to connect to data source',
               }
             )}
-            value={this.state.auth.credentials.username || ''}
+            value={this.state.auth.credentials?.username || ''}
             isInvalid={!!this.state.formErrorsByField.createCredential?.username?.length}
             onChange={this.onChangeUsername}
             onBlur={this.validateUsername}
@@ -625,19 +883,19 @@ export class EditDataSourceForm extends React.Component<EditDataSourceProps, Edi
                 )}
                 type={'dual'}
                 value={
-                  this.props.existingDataSource.auth.type !== AuthType.NoAuth
-                    ? '********'
-                    : this.state.auth.credentials.password
+                  this.props.existingDataSource.auth.type === AuthType.UsernamePasswordType
+                    ? this.maskedPassword
+                    : this.state.auth.credentials?.password
                 }
                 isInvalid={!!this.state.formErrorsByField.createCredential?.password?.length}
                 spellCheck={false}
                 onChange={this.onChangePassword}
                 onBlur={this.validatePassword}
-                disabled={this.props.existingDataSource.auth.type !== AuthType.NoAuth}
+                disabled={this.props.existingDataSource.auth.type === AuthType.UsernamePasswordType}
                 data-test-subj="updateDataSourceFormPasswordField"
               />
             </EuiFlexItem>
-            {this.props.existingDataSource.auth.type !== AuthType.NoAuth ? (
+            {this.props.existingDataSource.auth.type === AuthType.UsernamePasswordType ? (
               <EuiFlexItem>{this.renderUpdatePasswordModal()}</EuiFlexItem>
             ) : null}
           </EuiFlexGroup>
@@ -659,12 +917,17 @@ export class EditDataSourceForm extends React.Component<EditDataSourceProps, Edi
       auth.type === formValues.auth.type &&
       auth.type === AuthType.UsernamePasswordType &&
       formValues.auth.credentials?.username !== auth.credentials?.username;
+    const isRegionChanged =
+      auth.type === formValues.auth.type &&
+      auth.type === AuthType.SigV4 &&
+      formValues.auth.credentials?.region !== auth.credentials?.region;
 
     if (
       formValues.title !== title ||
       formValues.description !== description ||
       formValues.auth.type !== auth.type ||
-      isUsernameChanged
+      isUsernameChanged ||
+      isRegionChanged
     ) {
       this.setState({ showUpdateOptions: true });
     } else {